+++ examples/project/dialog_box/test_1_base_box.py
@@ -1,0 +1,27 @@
+import pytest
+
+from hardpy import run_dialog_box
+from hardpy.pytest_hardpy.utils import DialogBox
+
+
+pytestmark = pytest.mark.module_name("Base dialog box")
+
+
+@pytest.mark.case_name("Empty test before")
+def test_before():
+    assert True
+
+
+@pytest.mark.case_name("Base dialog box")
+def test_base_dialog_box():
+    dbx = DialogBox(
+        title_bar="Operator check",
+        dialog_text="Press the Confirm button",
+    )
+    response = run_dialog_box(dbx)
+    assert response is None
+
+
+@pytest.mark.case_name("Empty test after")
+def test_after():
+    assert True