+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add base dialog box, text input and numeric input.
+* Add dialog box invocation functionality.
+* Add a socket mechanism to transfer data from the uvicorn server to the pytest subprocess.
+
 ## HardPy 0.3.0
 
 * Add implementation of test dependencies without using third party plugins.